third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/response-json-expected.txt

This is a testharness.js-based test.
PASS json response with no data: response property is null
PASS json response with malformed data: response property is null
FAIL JSON object roundtripassert_equals: Response should be cached expected object "[object Object]" but got object "[object Object]"
FAIL JSON roundtrip with Japanese textassert_equals: Response should be cached expected object "[object Object]" but got object "[object Object]"
Harness: the test ran to completion.
